minor changes, mainly diagnostics for sync

+++ src/Nethermind/Nethermind.Network/P2P/Subprotocols/Eth/StatusMessage.cs
@@ -25,7 +25,7 @@ namespace Nethermind.Network.P2P.Subprotocols.Eth
     {
         public override int PacketType { get; } = 0;
         public override string Protocol { get; } = "eth";
-        public int ProtocolVersion { get; set; }
+        public byte ProtocolVersion { get; set; }
         public long NetworkId { get; set; }
         public BigInteger TotalDifficulty { get; set; }
         public Keccak BestHash { get; set; }
